added a page for about info that the user will be taken to after registering
