+++ src/components/auth/Auth.css
@@ -1,7 +1,7 @@
 /* Login Page */
 
 
-.login-page-container {
+.login-page-container, .register-form-container {
     background-image: url('./Lauren_v4.png');
     background-size: cover;
     background-repeat: no-repeat;
@@ -9,14 +9,7 @@
     text-align: center;
     opacity: .9;
 }
-.register-form-container {
-    background-image: url('./Lauren_v4.png');
-    background-size: cover;
-    background-repeat: no-repeat;
-    /* height: 100vh; */
-    text-align: center;
-    opacity: .9;
-}
+
 
 .login-form input, .register-form input, .register-form select {
     margin: 0 auto;
@@ -37,10 +30,10 @@
     margin-top: 20px;
 }
 
-.login-form button {
+.login-form button, .register-form button {
     background-color: #FEDA6A;
     border-color: #FEDA6A;
-    margin-bottom: 20px;
+    margin-bottom: 15px;
 }
 
 
@@ -60,20 +53,34 @@
     color: #FEDA6A;
 }
 
-.login-page-container a {
+.login-page-container a, .register-form-container a {
     color: #FEDA6A;
     margin-top: 40px;
     font-weight: bold;
 }
 
-.login-page-container a:hover {
+.login-page-container a:hover, .register-form-container a:hover{
     color: white;
     margin-top: 40px;
     font-weight: bold;
     text-decoration: none;
+    margin-bottom: 20px;
 }
-
-
+.register-link {
+    margin-bottom: 80px;
+}
 /* Register Page */
 
+@media screen and (max-width: 900px) {
+    
+    .register-form-container {
+        background-image: url('./Lauren_v4.png');
+        background-size: cover;
+        background-repeat: no-repeat;
+        text-align: center;
+        opacity: .9;
+
+    }
+    
+}
