+++ src/components/auth/Register.js
@@ -8,6 +8,8 @@ import './Auth.css'
 import { Button, FormControl, Form } from "react-bootstrap"
 import { cloudName, uploadPreset } from '../../modules/Credentials';
 import "../profile/Profile.css"
+import Logo from "../home/Hive_Loge.png"
+
 
 class Register extends Component {
 
@@ -16,17 +18,7 @@ class Register extends Component {
         userName: "",
         lastName: "",
         password: "",
-        firstName: "",
-        cohort: "",
-        cohorts: [],
-        employmentStatus: "",
-        linkedInProfile: "",
-        confirmpassword: "",
-        aboutme: "",
-        employer: "",
-        location: "",
-        jobtitle: "",
-        image: "https://freesvg.org/img/icon_user_whiteongrey.png"
+        firstName: ""
     }
 
     handleInputChange = (evt) => {
@@ -35,12 +27,7 @@ class Register extends Component {
         this.setState(stateToChange)
     }
 
-    componentDidMount() {
-        // get all cohorts to populate the dropdown
-        APIManager.getAllNotAuth("cohorts")
-            .then((cohorts) => this.setState({ cohorts: cohorts }))
-    }
-    handleRegister = event => {
+    pushToAbout = event => {
         event.preventDefault()
 
         // Create object with values from state
@@ -49,32 +36,17 @@ class Register extends Component {
             "first_name": this.state.firstName,
             "last_name": this.state.lastName,
             "email": this.state.email,
-            "password": this.state.password,
-            "cohort_id": this.state.cohort,
-            "is_employed": this.state.employmentStatus,
-            "linkedin_profile": this.state.linkedInProfile,
-            "aboutme": this.state.aboutme,
-            "image": this.state.image,
-            "employer": this.state.employer,
-            "location": this.state.location,
-            "jobtitle": this.state.jobtitle
-
+            "password": this.state.password
         }
 
         // check to see that all fields are filled out
-        if (this.state.password === this.state.confirmpassword && this.state.userName !== "" && this.state.firstName !== "" && this.state.lastName !== "" && this.state.email !== "" && this.state.password !== "" && this.state.confirmpassword !== "" && this.state.cohort !== "" && this.state.employmentStatus !== "") {
+        if (this.state.password === this.state.confirmpassword && this.state.userName !== "" && this.state.firstName !== "" && this.state.lastName !== "" && this.state.email !== "" && this.state.password !== "" && this.state.confirmpassword !== "") {
 
             // Make a fetch call with the object as the body of the POST request
-            register(newUser)
-                .then(() => {
-                    // checks to make sure the user is authenticated
-                    if (isAuthenticated()) {
-                        // this function sets the user value in state in hivemind.js to true so that the navbar will render
-                        this.props.loggedIn()
-                        // push to the home page
-                        this.props.history.push("/")
-                    }
-                })
+            this.props.history.push({
+                pathname: "/about",
+                state: { user: newUser }
+            })
         } else {
             // these alerts will be triggered if a field is not filled out or the passwords do not match 
             if (this.state.userName === "") {
@@ -85,10 +57,6 @@ class Register extends Component {
                 alert('Please provide a last name.')
             } else if (this.state.email === "") {
                 alert('Please provide an email address.')
-            } else if (this.state.cohort === "") {
-                alert('Please select a cohort.')
-            } else if (this.state.employmentStatus === "") {
-                alert('Please select an employment status.')
             } else if (this.state.password === "") {
                 alert('Please provide a password.')
             } else if (this.state.confirmpassword === "") {
@@ -100,136 +68,71 @@ class Register extends Component {
         }
     }
 
-    openCloudinaryWidget = () => {
-        let widget = window.cloudinary.createUploadWidget({
-            cloudName: cloudName,
-            uploadPreset: uploadPreset
-        }, (error, result) => {
-            if (!error && result && result.event === "success") {
-                this.setState({
-                    image: result.info.url
-                })
-            }
-        }
-        )
-        widget.open();
-    }
-
 
     render() {
 
         return (
-            <section className="register-form-container">
+            <>
+                <section className="register-form-container">
+                    <center><img src={Logo} alt="logo" className="home-logo" width="300" height="300" ></img></center>
                     <h1 className="register-header">Register for Hive Mind!</h1>
-                <Form className="register-form" onSubmit={this.handleRegister}>
-                    <FormControl
-                        id="userName"
-                        onChange={this.handleInputChange}
-                        placeholder="Username"
-                    />
+                    <Form className="register-form" onSubmit={this.pushToAbout}>
+                        <FormControl
+                            id="userName"
+                            onChange={this.handleInputChange}
+                            placeholder="Username"
+                        />
 
 
 
-                    <FormControl
-                        id="firstName"
-                        onChange={this.handleInputChange}
-                        placeholder="First Name"
+                        <FormControl
+                            id="firstName"
+                            onChange={this.handleInputChange}
+                            placeholder="First Name"
 
-                    />
+                        />
 
 
-                    <FormControl
-                        id="lastName"
-                        onChange={this.handleInputChange}
-                        placeholder="Last Name"
+                        <FormControl
+                            id="lastName"
+                            onChange={this.handleInputChange}
+                            placeholder="Last Name"
 
-                    />
+                        />
 
-                    <FormControl
-                        id="email"
-                        onChange={this.handleInputChange}
-                        placeholder="Email"
+                        <FormControl
+                            id="email"
+                            onChange={this.handleInputChange}
+                            placeholder="Email"
 
-                    />
-                    <select id="cohort"
-                        onChange={this.handleInputChange}>
+                        />
 
-                        <option value="">Select a Cohort</option>
-                        {this.state.cohorts.map((cohort) => {
-                            return <option key={cohort.id} value={cohort.id}>{cohort.cohort}</option>
-                        })}
-                    </select>
+                        <FormControl
+                            id="password"
+                            onChange={this.handleInputChange}
+                            type="password"
+                            placeholder="Password"
 
-                    <select id="employmentStatus"
-                        onChange={this.handleInputChange}>
-                        <option value="">Select Employment Status</option>
-                        <option value="True">Hired</option>
-                        <option value="False">Searching for Opportunities</option>
-                    </select>
+                        />
 
-                    <FormControl
-                        id="employer"
-                        onChange={this.handleInputChange}
-                        placeholder="If you are working somewhere, where are you employed?"
+                        <FormControl
+                            id="confirmpassword"
+                            onChange={this.handleInputChange}
+                            type="password"
+                            placeholder="Confirmation Password"
 
-                    />
-                    <FormControl
-                        id="jobtitle"
-                        onChange={this.handleInputChange}
-                        placeholder="What is your job title?"
+                        />
 
-                    />
-                    <FormControl
-                        id="linkedInProfile"
-                        onChange={this.handleInputChange}
-                        placeholder="LinkedIn Profile"
 
-                    />
-                    <FormControl
-                        id="aboutme"
-                        onChange={this.handleInputChange}
-                        placeholder="Tell everyone a little bit about yourself..."
-
-                    />
-                    <FormControl
-                        id="location"
-                        onChange={this.handleInputChange}
-                        placeholder="Tell us where your located..."
-
-                    />
-
-                    <FormControl
-                        id="password"
-                        onChange={this.handleInputChange}
-                        type="password"
-                        placeholder="Password"
-
-                    />
-
-                    <FormControl
-                        id="confirmpassword"
-                        onChange={this.handleInputChange}
-                        type="password"
-                        placeholder="Confirmation Password"
-
-                    />
-
-                    {this.state.image !== null &&
-                        <img src={this.state.image} alt="user" className="pre-profile-img"></img>
-                    }
-
-                    <div className="upload_widget_container">
-                        <Button  id="upload_widget" variant="secondary" onClick={this.openCloudinaryWidget}>Choose File</Button>
-                    </div>
-
-                    <Button variant="secondary" type="submit">
-                        Register
+                        <Button variant="secondary" type="submit">
+                            Next
                     </Button>
-                </Form>
-                <Link to="/login">Already have an account? Click here to sign in!</Link>
-            </section>
+                    </Form>
+                <Link clasName="register-link" to="/login">Already have an account? Click here to sign in!</Link>
+                </section>
+            </>
         )
     }
 }
 
-export default Register
+export default Register;
